draft: style macro build implementation
