+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["stylers_core", "stylers", "stylers_build", "stylers_test"]
+members = ["stylers_core", "stylers", "stylers_macro", "stylers_test"]
 exclude = ["examples"]
 
 [workspace.package]
