+++ examples/style_build/Cargo.toml
@@ -6,8 +6,9 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-leptos = { git = "https://github.com/leptos-rs/leptos.git" }
+leptos = { version = "0.4.9" }
 stylers = { path = "../../stylers" }
+gloo = "0.8.0"
 
 [build-dependencies]
-stylers_build = { path = "../../stylers_build" }
+stylers = { path = "../../stylers" }
