+++ stylers/Cargo.toml
@@ -8,11 +8,9 @@ repository = "https://github.com/abishekatp/stylers"
 description = "Scoped CSS for Rust web frameworks like leptos"
 readme = "../README.md"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[lib]
-proc_macro = true
-
 [dependencies]
-quote = "1.0"
-proc-macro2 = { version = "1.0" }
+glob = "0.3.1"
+syn = { version = "2.0.15", features = ["full"] }
 stylers_core = { path = "../stylers_core" }
+stylers_macro = { path = "../stylers_macro" }
+proc-macro2 = "1.0.60"
