+++ stylers_core/src/style/css_style_rule.rs
@@ -70,7 +70,7 @@ impl CSSStyleRule {
                             if ch == '.' || ch == '#' || ch == ':' {
                                 add_spaces(&mut selector, t.span(), &mut pre_line, &mut pre_col);
                             } else {
-                                let end = t.span().unwrap().end();
+                                let end = t.span().end();
                                 pre_col = end.column;
                                 pre_line = end.line;
                             }
